chore: reorganize install
