+++ src/core/client/install/App/App.css
