+++ src/core/client/install/App/App.spec.tsx
