+++ src/core/client/install/App/MainBar.css
