+++ src/core/client/install/App/MainBar.tsx
