+++ src/core/client/install/App/Wizard.css
