+++ src/core/client/install/App/index.ts
@@ -1,0 +1,1 @@
+export { default, default as App } from "./App";
