+++ src/core/client/install/App/steps/FinalStep.tsx
