+++ src/core/client/install/App/steps/InitialStep.tsx
