+++ src/core/client/install/App/steps/NextButton.tsx
