--- src/core/client/install/mutations/index.ts
@@ -1,1 +1,0 @@
-export { withInstallMutation, InstallMutation } from "./InstallMutation";
